+++ travis-install.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-curl -sL https://github.com/xd009642/tarpaulin/releases/download/0.13.0/cargo-tarpaulin-0.13.0-travis.tar.gz | tar xvz -C $HOME/.cargo/bin
+curl -sL https://github.com/xd009642/tarpaulin/releases/download/0.13.1/cargo-tarpaulin-0.13.1-travis.tar.gz | tar xvz -C $HOME/.cargo/bin